классы TrackAdapter и TrackViewHolder вынесены в отдельные файлы
